last hope

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@
 int main(void)
 {
 	/* printing size of datatypes */
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-	printf("Size of an int: 4 byte(s)\n", sizeof(int));
-	printf("Size of a long int: 4 byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: 8 byte(s)\n", sizeof(long long int));
-	printf("Size of a float: 4 byte(s)\n", sizeof(float));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	printf("Size of an int: 4 byte(s)\n", (unsigned long)sizeof(int));
+	printf("Size of a long int: 4 byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of a long long int: 8 byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of a float: 4 byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
 }
